Update gitlab.json and gitlab_connector.py
In gitlab_connectory.py, modify the action_result data for the list_triggers, list_users, list_branches, and list_projects actions so the returned data is not a nested array. Instead, the data returned is an array of the corresponding object type (project, user, branch, and trigger). Also update gitlab.json with the updated output values.